minor refactorings
