+++ mm2src/kdf_walletconnect/src/client/inbound.rs
@@ -1,8 +1,5 @@
 use crate::error::ClientError;
 
-use futures::SinkExt;
-use futures::TryFutureExt;
-use pin_project::pin_project;
 use relay_rpc::domain::MessageId;
 use relay_rpc::rpc::{self, ErrorResponse, Payload, Response, ServiceRequest, SuccessfulResponse};
 use tokio::sync::mpsc::UnboundedSender;
